requirements.txt
pytest
requests
pydantic
allure-pytest
